requirements.txt
boto3
argparse
python-dotenv